requirements.txt
beautifulsoup4==4.11.1
google-cloud-texttospeech==2.11.0